Move PaymentSheetViewModel#isGooglePayReady to SheetViewModel
This makes it reusable in `PaymentOptionsViewModel`

Only update `isGooglePayReady` if needed.
