+++ stripe/src/main/java/com/stripe/android/paymentsheet/PaymentSheetViewModel.kt
@@ -32,8 +32,6 @@ import com.stripe.android.paymentsheet.ui.SheetMode
 import com.stripe.android.paymentsheet.viewmodels.SheetViewModel
 import com.stripe.android.view.AuthActivityStarter
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.filterNotNull
-import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import kotlin.coroutines.CoroutineContext
@@ -43,19 +41,18 @@ internal class PaymentSheetViewModel internal constructor(
     private val stripeAccountId: String?,
     private val stripeRepository: StripeRepository,
     private val paymentController: PaymentController,
-    private val googlePayRepository: GooglePayRepository,
+    googlePayRepository: GooglePayRepository,
     internal val args: PaymentSheetActivityStarter.Args,
-    private val workContext: CoroutineContext
+    workContext: CoroutineContext
 ) : SheetViewModel<PaymentSheetViewModel.TransitionTarget, ViewState>(
-    isGuestMode = args is PaymentSheetActivityStarter.Args.Guest
+    isGuestMode = args is PaymentSheetActivityStarter.Args.Guest,
+    googlePayRepository = googlePayRepository,
+    workContext = workContext
 ) {
     private val confirmParamsFactory = ConfirmParamsFactory()
 
     private val mutablePaymentIntent = MutableLiveData<PaymentIntent?>()
     internal val paymentIntent: LiveData<PaymentIntent?> = mutablePaymentIntent
-
-    private val mutableIsGooglePayReady = MutableLiveData<Boolean>()
-    internal val isGooglePayReady: LiveData<Boolean> = mutableIsGooglePayReady.distinctUntilChanged()
 
     fun fetchAddPaymentMethodConfig() = liveData {
         emitSource(
@@ -211,16 +208,6 @@ internal class PaymentSheetViewModel internal constructor(
             }
             else -> {
                 // TODO(mshafrir-stripe): handle error
-            }
-        }
-    }
-
-    fun fetchIsGooglePayReady() {
-        viewModelScope.launch {
-            withContext(workContext) {
-                mutableIsGooglePayReady.postValue(
-                    googlePayRepository.isReady().filterNotNull().first()
-                )
             }
         }
     }
